Make item handles smaller
If someone has more than 65,535 items in one screen area at one time,
they are probably making a piece of software so complex that no one can
use it.

+++ include/wima.h
@@ -286,7 +286,7 @@ typedef struct wima_area_handle_struct {
  */
 typedef struct wima_item_handle {
 
-	int32_t item;
+	uint16_t item;
 
 	WimaAreaNodeHandle area;
 	WimaWindowHandle window;
